relayburn-cli: share HOME mutation lock across harness test modules
CodeRabbit + Codex flagged: codex.rs and opencode.rs each defined a
private `static ENV_LOCK: Mutex<()>` for serializing `HOME` mutation
in tests, but `HOME` is process-global. `cargo test` runs modules in
parallel by default, so the two modules' `with_test_home` helpers
could acquire their own locks simultaneously and race on the global
env, intermittently leaking each other's sentinel values into
`session_root()` assertions.

Extract the lock + helper into a shared `harnesses::test_env` module
gated by `#[cfg(test)]`. Both adapter test modules now `use
crate::harnesses::test_env::with_test_home;` and funnel through one
crate-wide `ENV_LOCK`. The save-set-restore-`resume_unwind`
machinery (including `catch_unwind` so a panicking assertion still
restores `HOME` before propagating) lives in one place — adapters
that join later (e.g. a future `aider` harness) reuse the helper
without re-deriving the unwind-safety details.

Scope kept tight: SDK-side `RELAYBURN_*` env tests in
`relayburn_sdk::query_verbs` carry their own lock and don't share
this one. Lifting both into a workspace-wide test-utility crate is a
deliberate follow-up if HOME and RELAYBURN_HOME mutations ever need
to interlock; today the two scopes are independent.
